corrigindo tipagem de nota

+++ src/main/java/com/education/education/dto/NotaResponseDTO.java
@@ -1,6 +1,7 @@
 package com.education.education.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record NotaResponseDTO(Integer nota, LocalDate data_lancamento, DisciplinaResponseDTO disciplina) {
+public record NotaResponseDTO(BigDecimal nota, LocalDate data_lancamento, DisciplinaResponseDTO disciplina) {
 }
